requirements.txt
asn1crypto==1.5.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
distlib==0.3.6
filelock==3.12.0
greenlet==2.0.2
idna==3.4
isort==5.12.0
mypy-extensions==1.0.0
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.2.0
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7
SQLAlchemy==1.4.47
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.22.0
